Project file is added here.
The MainActivity.java is the runnable class, however to execute this you need to create a new Android project and put this in the src\main\java subdirectory.  This was created using basic Android Studio tools to instantiate objects.  The goal is to get as far in the game as possible.